doc: discourage the use of nflog_set_nlbufsiz()
The default buffer size (which is one memory page) provides the
best results in terms of performance. For that reason, noone should
use nflog_set_nlbufsiz() to set up a higher buffer in the
applications (even if this is supported by the nfnetlink_log).

Davem spotted this during the Netfilter Workshop 2010.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>